moved cursor_on_closing_char_pair to line.py
