+++ bpython/curtsiesfrontend/manual_readline.py
@@ -5,8 +5,9 @@ and the cursor location
 based on http://www.bigsmoke.us/readline/shortcuts"""
 
 from ..lazyre import LazyReCompile
-from . import CHARACTER_PAIR_MAP
 import inspect
+
+from ..line import cursor_on_closing_char_pair
 
 INDENT = 4
 
@@ -229,28 +230,6 @@ def back_word(cursor_offset, line):
 @edit_keys.on("<DELETE>")
 def delete(cursor_offset, line):
     return (cursor_offset, line[:cursor_offset] + line[cursor_offset + 1 :])
-
-
-def cursor_on_closing_char_pair(cursor_offset, line):
-    """Checks if cursor sits on closing character of a pair
-    and whether its pair character is directly behind it
-    """
-    on_closing_char, pair_close = False, False
-    if line is None:
-        return on_closing_char, pair_close
-    if cursor_offset < len(line):
-        cur_char = line[cursor_offset]
-        if cur_char in CHARACTER_PAIR_MAP.values():
-            on_closing_char = True
-        if cursor_offset > 0:
-            prev_char = line[cursor_offset - 1]
-            if (
-                on_closing_char
-                and prev_char in CHARACTER_PAIR_MAP
-                and CHARACTER_PAIR_MAP[prev_char] == cur_char
-            ):
-                pair_close = True
-    return on_closing_char, pair_close
 
 
 @edit_keys.on("<BACKSPACE>")
